# perf(es/minifier): Make dead branch remover parallel