… crates/swc_ecma_minifier/tests/projects/output/angular-1.2.5.js